PB NH - MF-273 - Decoupled validation tests for telephone number
