+++ src/patient-registration/input/basic-input/telephone-number-input/telephone-number-input.component.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { useField } from 'formik';
-import styles from './../input.css';
+import styles from './../../input.css';
 
 interface TelephoneNumberInputProps {
   label: string;
