+++ src/patient-registration/patient-registration-validation.tsx
@@ -1,0 +1,16 @@
+import * as Yup from 'yup';
+
+export const validationSchema = Yup.object({
+  givenName: Yup.string().required('Given name is required'),
+  familyName: Yup.string().required('Family name is required'),
+  gender: Yup.string()
+    .oneOf(['M', 'F', 'O', 'U'], 'Gender is unspecified')
+    .required('Gender is required'),
+  birthdate: Yup.date()
+    .required('Birthdate is required')
+    .max(Date(), 'Birthdate cannot be in the future')
+    .nullable(),
+  yearsEstimated: Yup.number().min(0, 'Years cannot be less than 0'),
+  monthsEstimated: Yup.number().min(0, 'Months cannot be less than 0'),
+  telephoneNumber: Yup.string().matches(/^[0-9]*$/, 'Telephone number should only contain digits'),
+});
